+++ HISTORY.rst
@@ -1,3 +1,13 @@
+0.0.3 (2015-03-12)
+++++++++++++++++++
+* Remove ``user`` setting
+* Remove dependency lib [gistapi]
+* Enhancement for gist selection of ``open gist``
+* Add a new module ``api.py``
+* Add ``reload gist workspace`` command
+* Add ``reload gist cache`` command
+
+
 0.0.2 (2015-03-12)
 ++++++++++++++++++
 * Add ``update gist`` command